# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VeilGraph RightsGate",
    "description": "Evidence-first pre-publication media trust gateway derived from VeilGraph. The API exposes the inherited privacy foundation plus versioned RightsGate contracts, offline C2PA inspection, governed image retrieval, licence evaluation, deterministic policy and durable non-authorizing image assessments.",
    "version": "0.2.0-rightsgate-prototype"
  },
  "paths": {
...
    "/api/v1/rightsgate/requests/validate": {
      "post": {
        "tags": [
          "RightsGate"
        ],
        "summary": "Validate Request",
        "description": "Validate and fingerprint an idempotent assessment request envelope.",
        "operationId": "validate_request_api_v1_rightsgate_requests_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RightsGateAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestValidationReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rightsgate/assessments/validate": {
      "post": {
        "tags": [
          "RightsGate"
        ],
        "summary": "Validate Assessment",
        "description": "Validate an externally assembled result without granting release authority.",
        "operationId": "validate_assessment_api_v1_rightsgate_assessments_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RightsGateAssessment-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentValidationReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rightsgate/provenance/c2pa": {
      "post": {
        "tags": [
          "RightsGate"
        ],
        "summary": "Inspect C2Pa",
        "description": "Run the bounded offline C2PA adapter without making a release decision.",
        "operationId": "inspect_c2pa_api_v1_rightsgate_provenance_c2pa_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_inspect_c2pa_api_v1_rightsgate_provenance_c2pa_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2PAVerificationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rightsgate/rights/references/image": {
      "post": {
        "tags": [
          "RightsGate"
        ],
        "summary": "Derive Image Reference",
        "description": "Derive a byte-bound governed reference record using the server adapter.",
        "operationId": "derive_image_reference_api_v1_rightsgate_rights_references_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_derive_image_reference_api_v1_rightsgate_rights_references_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryImage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rightsgate/assessments": {
      "post": {
        "tags": [
          "RightsGate"
        ],
        "summary": "Execute Assessment",
        "description": "Run the implemented lanes under durable idempotency and fail-closed policy.",
        "operationId": "execute_assessment_api_v1_rightsgate_assessments_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_execute_assessment_api_v1_rightsgate_assessments_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentExecutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rightsgate/assessments/{idempotency_key}": {
      "get": {
        "tags": [
          "RightsGate"
        ],
        "summary": "Get Assessment Record",
        "description": "Retrieve durable status and a completed result without raw asset bytes.",
        "operationId": "get_assessment_record_api_v1_rightsgate_assessments__idempotency_key__get",
        "parameters": [
          {
            "name": "idempotency_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssessmentExecutionResponse": {
        "properties": {
          "receipt_schema": {
            "type": "string",
            "const": "veilgraph.rightsgate.execution-receipt.v1",
            "title": "Receipt Schema",
            "default": "veilgraph.rightsgate.execution-receipt.v1"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "request_sha256": {
            "type": "string",
            "title": "Request Sha256"
          },
          "execution_sha256": {
            "type": "string",
            "title": "Execution Sha256"
          },
          "assessment_sha256": {
            "type": "string",
            "title": "Assessment Sha256"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed"
          },
          "release_authorization": {
            "type": "boolean",
            "const": false,
            "title": "Release Authorization",
            "default": false
          },
          "assessment": {
            "$ref": "#/components/schemas/RightsGateAssessment-Output"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "request_sha256",
          "execution_sha256",
          "assessment_sha256",
          "replayed",
          "assessment"
        ],
        "title": "AssessmentExecutionResponse"
      },
...
      "AssessmentValidationReceipt": {
        "properties": {
          "receipt_schema": {
            "type": "string",
            "const": "veilgraph.rightsgate.assessment-validation.v1",
            "title": "Receipt Schema",
            "default": "veilgraph.rightsgate.assessment-validation.v1"
          },
          "accepted_schema": {
            "type": "string",
            "const": "veilgraph.rightsgate.assessment.v1",
            "title": "Accepted Schema",
            "default": "veilgraph.rightsgate.assessment.v1"
          },
          "assessment_id": {
            "type": "string",
            "title": "Assessment Id"
          },
          "assessment_sha256": {
            "type": "string",
            "title": "Assessment Sha256"
          },
          "deployment_decision": {
            "$ref": "#/components/schemas/DeploymentDecision"
          },
          "release_authorization": {
            "type": "boolean",
            "const": false,
            "title": "Release Authorization",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "assessment_id",
          "assessment_sha256",
          "deployment_decision"
        ],
        "title": "AssessmentValidationReceipt"
      },
...
      "Body_derive_image_reference_api_v1_rightsgate_rights_references_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "reference_id": {
            "type": "string",
            "title": "Reference Id"
          },
          "kind": {
            "$ref": "#/components/schemas/ReferenceKind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "rights_holder": {
            "type": "string",
            "title": "Rights Holder"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          }
        },
        "type": "object",
        "required": [
          "file",
          "reference_id",
          "kind",
          "title",
          "rights_holder",
          "source_record_id"
        ],
        "title": "Body_derive_image_reference_api_v1_rightsgate_rights_references_image_post"
      },
      "Body_execute_assessment_api_v1_rightsgate_assessments_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "request_json": {
            "type": "string",
            "title": "Request Json"
          },
          "rights_registry_json": {
            "type": "string",
            "title": "Rights Registry Json"
          },
          "licence_registry_json": {
            "type": "string",
            "title": "Licence Registry Json"
          },
          "policy_json": {
            "type": "string",
            "title": "Policy Json"
          },
          "max_hamming_distance": {
            "type": "integer",
            "title": "Max Hamming Distance",
            "default": 6
          }
        },
        "type": "object",
        "required": [
          "file",
          "request_json",
          "rights_registry_json",
          "licence_registry_json",
          "policy_json"
        ],
        "title": "Body_execute_assessment_api_v1_rightsgate_assessments_post"
      },
      "Body_inspect_c2pa_api_v1_rightsgate_provenance_c2pa_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_inspect_c2pa_api_v1_rightsgate_provenance_c2pa_post"
      },
...
      "C2PAVerificationResult": {
        "properties": {
          "component": {
            "$ref": "#/components/schemas/ComponentRecord"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidencePointer"
            },
            "type": "array",
            "title": "Evidence"
          },
          "assessment": {
            "$ref": "#/components/schemas/ProvenanceAssessment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "component",
          "evidence",
          "assessment"
        ],
        "title": "C2PAVerificationResult"
      },
...
      "ClaimDimension": {
        "type": "string",
        "enum": [
          "PROVENANCE",
          "RIGHTS_EXPOSURE",
          "DEPLOYMENT_READINESS"
        ],
        "title": "ClaimDimension"
      },
...
      "DeploymentDecision": {
        "type": "string",
        "enum": [
          "GO",
          "REVIEW",
          "BLOCK"
        ],
        "title": "DeploymentDecision"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReferenceKind": {
        "type": "string",
        "enum": [
          "COPYRIGHTED_WORK",
          "TRADEMARK"
        ],
        "title": "ReferenceKind"
      },
      "RegistryImage": {
        "properties": {
          "reference_id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Reference Id"
          },
          "kind": {
            "$ref": "#/components/schemas/ReferenceKind"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "rights_holder": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Rights Holder"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "dhash": {
            "type": "string",
            "pattern": "^[0-9a-f]{16}$",
            "title": "Dhash"
          },
          "media_type": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Media Type"
          },
          "source_record_id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Source Record Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reference_id",
          "kind",
          "title",
          "rights_holder",
          "sha256",
          "dhash",
          "media_type",
          "source_record_id"
        ],
        "title": "RegistryImage"
      },
...
      "RequestValidationReceipt": {
        "properties": {
          "receipt_schema": {
            "type": "string",
            "const": "veilgraph.rightsgate.request-validation.v1",
            "title": "Receipt Schema",
            "default": "veilgraph.rightsgate.request-validation.v1"
          },
          "accepted_schema": {
            "type": "string",
            "const": "veilgraph.rightsgate.assessment-request.v1",
            "title": "Accepted Schema",
            "default": "veilgraph.rightsgate.assessment-request.v1"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "request_sha256": {
            "type": "string",
            "title": "Request Sha256"
          },
          "assessment_started": {
            "type": "boolean",
            "const": false,
            "title": "Assessment Started",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "request_sha256"
        ],
        "title": "RequestValidationReceipt"
      },
...
      "RightsGateAssessment-Input": {
        "properties": {
          "schema": {
            "type": "string",
            "const": "veilgraph.rightsgate.assessment.v1",
            "title": "Schema",
            "default": "veilgraph.rightsgate.assessment.v1"
          },
          "assessment_id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Assessment Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "asset": {
            "$ref": "#/components/schemas/AssetDescriptor"
          },
          "context": {
            "$ref": "#/components/schemas/AssessmentContext"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ComponentRecord"
            },
            "type": "array",
            "minItems": 1,
            "title": "Components"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidencePointer"
            },
            "type": "array",
            "title": "Evidence"
          },
          "exposure_graph": {
            "$ref": "#/components/schemas/AssetExposureGraph-Input"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceAssessment"
          },
          "rights": {
            "$ref": "#/components/schemas/RightsAssessment"
          },
          "deployment": {
            "$ref": "#/components/schemas/DeploymentAssessment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "assessment_id",
          "created_at",
          "asset",
          "context",
          "components",
          "exposure_graph",
          "provenance",
          "rights",
          "deployment"
        ],
        "title": "RightsGateAssessment"
      },
...
      "RightsGateAssessmentRequest": {
        "properties": {
          "schema": {
            "type": "string",
            "const": "veilgraph.rightsgate.assessment-request.v1",
            "title": "Schema",
            "default": "veilgraph.rightsgate.assessment-request.v1"
          },
          "idempotency_key": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Idempotency Key"
          },
          "asset_ir": {
            "$ref": "#/components/schemas/AssetIR"
          },
          "context": {
            "$ref": "#/components/schemas/AssessmentContext"
          },
          "requested_dimensions": {
            "items": {
              "$ref": "#/components/schemas/ClaimDimension"
            },
            "type": "array",
            "title": "Requested Dimensions",
            "default": [
              "PROVENANCE",
              "RIGHTS_EXPOSURE",
              "DEPLOYMENT_READINESS"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "asset_ir",
          "context"
        ],
        "title": "RightsGateAssessmentRequest"
      },
...
    }
  }
}